GP2-2275: Rename 'js' directory to 'javascript' directory across project
This changeset moves anything from /js/ into /javascript/ to avoid clashes
with the (temporary) custom routing on deployed environments that allows
great-domestic-ui to still get its assets from /static/js/. (Basically if
both great-cms and great-domestic-ui are trying to get assets from /static/js/
only one of them is able to have the route point to their service - and
it has to be g-d-ui that wins because we don't want to change its code).

As such, moving great-cms's scripts into /javascript/ will avoid this problem.
