+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Enhancements
 - GP2-1253 - Export plan - Learning linking components
+
+- GP2-2275 - Rename 'js' directory to 'javascript' directory across project
 - GP2-2214 - Signed in magna header (BAU pages)
 - GP2-1887 - VFM questionnaire
 - GP2-2086 - Date picker selectable ranges
