+++ contact/templates/domestic/contact/step.html
@@ -17,7 +17,7 @@
 
 {% block body_js %}
     {{ block.super }}
-    <script type="text/javascript" src="{% static 'js/dit.components.toggleOther.js' %}"></script>
+    <script type="text/javascript" src="{% static 'javascript/dit.components.toggleOther.js' %}"></script>
     <script type="text/javascript">
         var companyTypeOther = document.getElementById('id_company_type_1');
         var otherInputs = document.getElementById('id_company_type_other-container');
