+++ core/components/static/javascript/dit.components.cookie-notice.js
