+++ core/components/static/javascript/dit.components.header.js
