+++ core/components/static/javascript/dit.responsive.js
