+++ core/components/static/javascript/reactComponents.js
