+++ core/static/javascript/auto-resize-textarea.js
