+++ domestic/static/javascript/marketFilters.js
