+++ domestic/templates/domestic/base.html
@@ -59,14 +59,14 @@
         </noscript>
 
         {% block vendor_js %}
-            <script src="{% static 'js/vendor/details-polyfill.js' %}"></script>
-            <script src="{% static 'js/vendor/jquery-3.5.1.min.js' %}"></script>
+            <script src="{% static 'javascript/vendor/details-polyfill.js' %}"></script>
+            <script src="{% static 'javascript/vendor/jquery-3.5.1.min.js' %}"></script>
         {% endblock %}
 
         {% block head_js %}
-            <script src="{% static 'js/dit.js' %}"></script>
-            {% comment "Disabled because cookie behaviour needs attention" %} <script src="{% static 'js/dit.components.cookie-notice.js' %}"></script> {% endcomment %}
-            {% comment "Disabled because cookie behaviour needs attention" %} <script src="{% static 'js/reactComponents.js' %}"></script> {% endcomment %}
+            <script src="{% static 'javascript/dit.js' %}"></script>
+            {% comment "Disabled because cookie behaviour needs attention" %} <script src="{% static 'javascript/dit.components.cookie-notice.js' %}"></script> {% endcomment %}
+            {% comment "Disabled because cookie behaviour needs attention" %} <script src="{% static 'javascript/reactComponents.js' %}"></script> {% endcomment %}
             <script>
               // Safari saves a snapshot of a page as it was when the user left, so upon going "back" it doesn’t have to reload and instead it loads the saved snapshot state. That has an unwanted side effect that cookie modal is shown when clicking "go back to the page you came from".
               window.onpageshow = function (event) {
@@ -78,7 +78,7 @@
         {% endblock %}
 
         {% block head_js_ga360 %}
-            <script src="{% static 'js/dit.tagging.js' %}"></script>
+            <script src="{% static 'javascript/dit.tagging.js' %}"></script>
             <script>
                 window.dataLayer = window.dataLayer || [];
                 {% if ga360 and not skip_ga360 %}
@@ -96,9 +96,9 @@
         {% endblock %}
 
         {% block header_js %}
-            <script src="{% static 'js/dit.responsive.js' %}"></script>
-            <script src="{% static 'js/dit.utils.js' %}"></script>
-            <script src="{% static 'js/dit.components.header.js' %}"></script>
+            <script src="{% static 'javascript/dit.responsive.js' %}"></script>
+            <script src="{% static 'javascript/dit.utils.js' %}"></script>
+            <script src="{% static 'javascript/dit.components.header.js' %}"></script>
         {% endblock %}
 
         {% block head_other %}
