+++ domestic/templates/domestic/country_guide.html
@@ -7,7 +7,7 @@
 
 {% block head_js %}
   {{ block.super }}
-  <script src="{% static 'js/country-guide.js' %}"></script>
+  <script src="{% static 'javascript/country-guide.js' %}"></script>
 {% endblock %}
 
 {% block content %}
