+++ domestic/templates/domestic/landing_page.html
@@ -10,9 +10,9 @@
 
 {% block head_js %}
   {{ block.super }}
-  <script src="{% static 'js/dit.classes.Modal.js' %}"></script>
-  <script src="{% static 'js/dit.components.video.js' %}"></script>
-  <script src="{% static 'js/home.js' %}"></script>
+  <script src="{% static 'javascript/dit.classes.Modal.js' %}"></script>
+  <script src="{% static 'javascript/dit.components.video.js' %}"></script>
+  <script src="{% static 'javascript/home.js' %}"></script>
 {% endblock %}
 
 {% block head_title %}{% firstof page.seo_title page.title %}{% endblock %}
