+++ domestic/templates/domestic/topic_landing_pages/markets.html
@@ -160,7 +160,7 @@
 
 {% block body_js %}
 {{ block.super }}
-<script src="{% static 'js/marketFilters.js' %}"></script>
+<script src="{% static 'javascript/marketFilters.js' %}"></script>
 <script>
   $(function() {
     marketFilters.init();
