+++ domestic/templates/domestic/ukef/contact_form.html
@@ -14,7 +14,7 @@
 
 {% block body_js %}
   {{ block.super }}
-  <script type="text/javascript" src="{% static 'js/dit.components.toggleOther.js' %}"></script>
+  <script type="text/javascript" src="{% static 'javascript/dit.components.toggleOther.js' %}"></script>
   <script type="text/javascript">
     var likeToDiscussInputYes = document.getElementById('id_like_to_discuss_1');
     var likeToDiscussOther = document.getElementById('id_like_to_discuss_other-container');
